se ha añadido el 3et ej de la ultima tarea

+++ Unidad-2/trabajos-metodos-2/src/main/java/ies/puerto/Ejercicio1.java
@@ -2,8 +2,8 @@ package ies.puerto;
 
 public class Ejercicio1 {
 
-    public int calcularSueldoHoras(int numeroHoras){
-        int sueldo = 0;
+    public float calcularSueldoHoras(float numeroHoras){
+        float sueldo = 0;
 
         sueldo = (numeroHoras*10);
